✨ Add check for duplicate files (#1)

+++ src/fstorage/server/storage/router.py
@@ -7,10 +7,10 @@ from fastapi.responses import FileResponse
 from sqlmodel import select
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from src.fstorage.server.app_params import APP_PARAMS
-
 from .exceptions import IntegrityError, NotFoundError
 from .models import Metadata
+
+from ..app_params import APP_PARAMS
 
 
 ROUTER = APIRouter(tags=['File storage'], prefix="/fstorage")
@@ -26,19 +26,29 @@ async def upload(file: UploadFile) -> dict[str, str]:
     :rtype: dict[str, str]
     """
     file_data: bytes = await file.read()
-    metadata = Metadata(
-        upload_timestamp=datetime.utcnow().timestamp(),
-        file_name=file.filename,
-        file_hash=hashlib.sha1(file_data).hexdigest(),
-        file_size=len(file_data)
-    )
-    response = {'id': str(metadata.id)}
+    file_hash: str = hashlib.sha1(file_data).hexdigest()
 
-    with open(APP_PARAMS.storage_path / f"{metadata.id}.bin", 'wb') as storage:
-        storage.write(file_data)
+    statement = select(Metadata).where(Metadata.file_hash == file_hash)
+    statement = statement.where(Metadata.file_size == len(file_data))
     async with AsyncSession(APP_PARAMS.db_engine) as session:
-        session.add(metadata)
-        await session.commit()
+        exist_metadata = (await session.execute(statement)).scalar()
+
+    if exist_metadata is None:
+        metadata = Metadata(
+            upload_timestamp=datetime.utcnow().timestamp(),
+            file_name=file.filename,
+            file_hash=file_hash,
+            file_size=len(file_data)
+        )
+        with open(APP_PARAMS.storage_path / f"{metadata.id}.bin", 'wb') as storage:
+            storage.write(file_data)
+
+        response = {'id': str(metadata.id)}
+        async with AsyncSession(APP_PARAMS.db_engine) as session:
+            session.add(metadata)
+            await session.commit()
+    else:
+        response = {'id': str(exist_metadata.id)}
 
     return response
 
@@ -56,11 +66,11 @@ async def get(file_uuid: str) -> FileResponse:
     filepath = APP_PARAMS.storage_path / f"{file_uuid}.bin"
     if not filepath.exists():
         raise NotFoundError
-
     with open(filepath, 'rb') as storage:
         file_data = storage.read()
+
+    statement = select(Metadata).where(Metadata.id == file_uuid)
     async with AsyncSession(APP_PARAMS.db_engine) as session:
-        statement = select(Metadata).where(Metadata.id == file_uuid)
         metadata = (await session.execute(statement)).scalar()
 
     # Check №2.
@@ -90,3 +100,4 @@ async def delete(file_uuid: str) -> None:
         statement = select(Metadata).where(Metadata.id == file_uuid)
         metadata = (await session.execute(statement)).scalar()
         await session.delete(metadata)
+        await session.commit()
